% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{calc_income_tax}
\alias{calc_income_tax}
\title{Given an income amount, standard deduction, tax brackets, and bracket minimums, calculates the
tax (dollar amount), tax bracket (percent), and effective tax rate (percent)}
\usage{
calc_income_tax(income, standard_deduction, tax_brackets, tax_brackets_mins)
}
\arguments{
\item{income}{The amount of income to calculate the tax for}

\item{standard_deduction}{The standard deduction for the tax type. Specify 0 if none.}

\item{tax_brackets}{A vector of the tax bracket percentages (specified as numbers from 0 - 100)}

\item{tax_brackets_mins}{A vector of the lower bound dollar amounts associated with each tax bracket}
}
\value{
A list of 3 numbers giving the tax (dollar amount), tax bracket (percent), and effective tax rate (percent)
}
\description{
Given an income amount, standard deduction, tax brackets, and bracket minimums, calculates the
tax (dollar amount), tax bracket (percent), and effective tax rate (percent)
}
