% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{calc_tax_table}
\alias{calc_tax_table}
\title{Compile the tax calculations into a tabular arrangement of line items for presentation}
\usage{
calc_tax_table(income_salary, income_401k, tax_config)
}
\arguments{
\item{income_salary}{The amount of income coming from a salary}

\item{income_401k}{The amount of income coming from a 401k distribution}

\item{tax_config}{A dictionary with the tax brackets and rates}
}
\value{
A DataFrame for presenting the tax calculation line items
}
\description{
Compile the tax calculations into a tabular arrangement of line items for presentation
}
